add in per file meta preview option
