+++ app/templates/embed/preview.html
@@ -13,7 +13,7 @@
 
     <meta property="og:author" content="{{ file.user.get_name }}">
     <meta property="og:title" content="{{ file.name }}">
-    <meta property="og:description" content="{{ file.info|default:"" }}{% if file.user.show_exif_preview %}{% if file.exif.Model %}
+    <meta property="og:description" content="{{ file.info|default:"" }}{% if file.meta_preview %}{% if file.exif.Model %}
 Camera: {% if file.exif.Make not in file.exif.Model %}{{ file.exif.Make }} {% endif %}{{ file.exif.Model }}{% endif %}{% if file.exif.DateTimeOriginal%}
 {{ file.exif.DateTimeOriginal|convert_str_date }}{% if file.meta.GPSArea %}{% endif %}{% if file.meta.GPSArea %}
 {{ file.meta.GPSArea }}{% endif %}{% if file.exif.GPSInfo.6 %}{% endif %}{% if file.exif.GPSInfo.6 %}
@@ -136,7 +136,7 @@ Resolution: {{ file.meta.PILImageWidth }} x {{ file.meta.PILImageHeight }}{% end
                     </p>
                     <p class="card-text mx-0 mb-0">Uploader: {{ file.user.get_name }}</p>
                     <p class="card-text">Uploaded {{ file.date }}</p>
-                    {% if file.user.show_exif_preview and file.exif %}
+                    {% if file.meta_preview and file.exif %}
                         <p class="card-text text-body-secondary">
                             {% if file.exif.DateTime %}<b>Taken On:</b> {{ file.exif.DateTimeOriginal|convert_str_date }}<br>{% endif %}
                             {% if file.exif.Model %}<b>Camera:</b> {% if file.exif.Make not in file.exif.Model %}{{ file.exif.Make }} {% endif %}{{ file.exif.Model }}<br>{% endif %}
